update unit tests for AppTable and ShowControls
